changed a lot
deleted jetpack code in the PlayerCharacterController and in UI. made it more general. the original player prefab will be the base of what I'll be using to remake some of the overwatch cast. the overwatch cast will be created off a prefab variant of the player prefab.

crouching only happens when grouded. you can't crouch when in air anymore. you have to hold crouch to stay crouched. can jump and hold crouch input to instantly crouch upon landing.

refactored the PlayerCharacterController because the  {} separation was more confusing and should've just put them in separate functions. Added names to it and changed it around. Still works the same besides the crouching.

added tracer script to temporarily figure out how to do her blink. so far not very successful. added some raytracing to measure the distance where i will blink. can't currently blink yet.